Add a multiplexing TCP front-end to tinydns
This uses the same engine as the existing UDP front-end, but can
furnish longer responses of up to 65535 bytes over a TCP transport.
Legacy AXFR/IXFR support is not included.

Keeping in mind the scalability limits of portable poll(), a pool of
up to 256 concurrent client streams is supported. If all slots are
filled, the longest idle connection is dropped to accept a new one.
However, multiple tcpdns instances can bind to the same addresses and
will load-balance using SO_REUSEPORT (or SO_REUSEPORT_LB on FreeBSD)
just as the existing UDP client does.

The new TCP daemon is tcpdns so rename the existing UDP daemon to udpdns
to match. This also clarifies that these daemons do not replicate the
interface of the original djbdns tinydns.