Respond to invalid queries in line with RFC 8906 advice
We already respond with RCODE_REFUSED to queries for which we have neither
authority nor a delegation, rather than silently dropping them like djbdns.

Respond to other invalid or unsupported queries as recommended in RFC 8906,
reworking response_query() and response_rcode() to make this convenient.
Ensure the Z and RA bits are clear in responses even when they were set
in the query.

As lookup() unconditionally calls response_rcode(), we need not set a
default RCODE_NOERROR or the QR bit in response_query() and can delegate
this job to response_rcode().