Use an unlinked temporary file for dnsdata -n
Allow cdb_make_start() to take a null filename which configures it
to operate in dummy mode and write to an unlinked temporary file from
tmpfile(). Use this to implement the -n option of dnsdata.

In dnsdata -n mode, report a short summary of program input and output
to stdout. The stderr output remains uncluttered and easy to parse for
pinpointing problem lines in data files.