Set RCODE_NOERROR when an alias chain strays out of bailiwick
This fixes a bug which led to invalid responses for CNAME chains whose
last entry points out of bailiwick: neither the rcode nor the QR bit was
set correctly before returning early.

Since we now need to call response_rcode(RCODE_NOERROR) in multiple places,
it is more economical to set it as the default in response_query() again.

Clarify the logic in response_finish(). If we have to truncate, there
are two cases: either we have a question and can fit it in the truncated
length, or we return just a header with QDCOUNT, ANCOUNT, NSCOUNT and
ARCOUNT all zero. We never need truncate smaller than the 12-byte header.