Support subtree mapping with synthetic CNAMEs
When a C line's target begins with "*.", serve synthetic CNAMEs in which
the "*." is replaced with the query name prefix that matched a wildcard
at the start of the line's label. Thus

  C*.example.org:*.example.com

will serve CNAMEs mapping every node under example.org to the corresponding
node under example.com.

This is the same functionality as DNAME but externally indistinguishable
from individual CNAMEs, sidestepping any issues with interoperability.

Avoid memcmp(s, "\1*", 2) when s might be the empty name "\0". This isn't
guaranteed to short-circuit and might read beyond the first byte to give
theoretically undefined behaviour. Fix the instance of this in dnsdata.c.
All our other memcmp() calls are fine as they compare fixed-sized arrays.