Drop the IP field from '.', '&', '@' and 'S' records
The IP field was a shorthand allowing an A or AAAA record to be created
for a hostname at the same time that hostname was used as the target of
an NS, MX or SRV record. For example,

  @example.com:1.2.3.4:mail.example.org:0

was equivalent to

  @example.com::mail.example.org:0
  +mail.example.org:1.2.3.4

In NS records, 'foo.ns.example.com' could be abbreviated to 'foo', and
similarly 'bar' stood for 'bar.mx.example.com' in MX records.

However, the shorthand added complexity and special cases to authorisation
checks and validation scripts. It was not sufficient to check if a user
is permitted to edit records for the direct label; it was also necessary
to check for a non-empty IP field and then check authorisation for adding
the nameserver host if this is not within the zone.

This extra work would be worthwhile if the benefits were sufficient, but
typical hosting providers have a small number of official name server and
mail server hostnames in their primary domain and use these in MX and NS
records for every zone, rather than maintaining many independent A and
AAAA records for the same machines in each zone. Even in their primary
domain, the NS and MX hostnames probably want reverse DNS, and there's
no way to specify '=' behaviour via the IP field.

Glueless delegation of non-primary domains is indeed less efficient, as
Dan correctly points out, but it is not as painful as maintaining large
numbers of glue records through painful out-of-band registry interfaces.
Our own data file has many thousands of '.' and '@' lines with a noisy
double colon after the domain name (invariably catching out users) and
only two places where the short notation could have been used naturally.

PTR records created by '=' lines also complicate validation: is the user
authorised to create a reverse record for the given address as well as the
directly labelled forward record? However, this is a worthwhile trade-off
in practice for brevity, readability and guaranteed consistency between
forward and reverse zones, especially with abbreviated IPv6 addresses.

This is a breaking change in the dnsdata input format. Check for
non-trivial use of the IP field with

  grep '^[.&@S][^:]*:[^:]' data

Once any instances have been replaced with separate '+' or '=' lines,
the remaining empty IP fields can be removed with

  sed -i 's/^\([.&@S][^:]*:\):/\1/' data

assuming no use of \: in data line labels.